전투 테스트씬 작업 / 스킬 작업 중 / 스킬 데이터 컬럼 수정 / 잘못 올린 데이터 삭제

+++ Assets/Scripts/Data/DBClass.cs
@@ -22,8 +22,8 @@ public class CharacterData : ReadData
     public int id;
     public int rarity;
     public string name;
-    public int attack;
-    public int defence;
+    public float attack;
+    public float defence;
     public int size;
     public int species;
     public int affiliation;
